refactor(test): move autocapture test activities to autocapture package
Move XmlAutocaptureTestActivity and ComposeAutocaptureTestActivity from
mpmetrics to autocapture package. Test classes remain in mpmetrics since
they use MixpanelAPI's package-private constructor.
