+++ analytics/src/androidTest/java/com/mixpanel/android/mpmetrics/ComposeAutocaptureInstrumentedTest.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.content.SharedPreferences
 import android.os.SystemClock
 import android.view.MotionEvent
+import com.mixpanel.android.autocapture.ComposeAutocaptureTestActivity
 import androidx.compose.ui.test.SemanticsNodeInteraction
 import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
